HW1/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.2
python-dotenv==1.1.1